PA: Convert metadata to pupa.
