+++ flake8.sh
@@ -2,5 +2,6 @@ python3 -m flake8 \
     openstates/fl \
     openstates/nc \
     openstates/mt \
+    openstates/pa \
     openstates/va \
     pupa2billy
